Fix CMakeLists
